HW3-FlightBooking/flight-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pyjwt>=2.8.0
